requirements.txt
ipykernel
pandas
firecrawl-py
bs4
python-dotenv
pytest-playwright
playwright
nest-asyncio
streamlit
